Updated FilesSystem Libraries
Updated FilesSystem Libraries

+++ Internal/Libraries/Helpers/Converter/Converter.php
@@ -15,12 +15,12 @@ class InternalConverter extends \CallController implements ConverterInterface
     // Byte
     //--------------------------------------------------------------------------------------------------------
     // 
-    // @param int  $bytes
-    // @param int  $precision
-    // @param bool $unit
+    // @param float $bytes
+    // @param int   $precision
+    // @param bool  $unit
     //
     //--------------------------------------------------------------------------------------------------------
-    public function byte(Int $bytes, Int $precision = 1, Bool $unit = true) : String
+    public function byte(Float $bytes, Int $precision = 1, Bool $unit = true) : String
     {       
         $byte   = 1024;
         $kb     = 1024 * $byte;
